refactor(core): align value-object naming, delete an orphaned example
Naming/placement consistency (no behaviour change):

- value_objects/role/mod.rs was the only VO family whose mod.rs did not
  re-export its type; add `pub use name::RoleName` and collapse the ~12
  verbose `role::name::RoleName` call sites to the conventional `role::RoleName`.
- Rename the two divergent name-VO leaf files to the `name.rs` convention
  used by the six sibling families (app/organization/project/secret/user/role):
  pipeline/pipeline_name.rs -> pipeline/name.rs, trigger/trigger_name.rs ->
  trigger/name.rs. The public paths (`pipeline::PipelineName`,
  `trigger::TriggerName`) are unchanged via the glob re-export.
- Delete examples/pipeline-verbose.json: orphaned (referenced nowhere) and
  on a pre-Step node schema the current API rejects.

Full workspace suite green, clippy clean.
