+++ crates/scylla-core/src/application/authz/grant.rs
@@ -6,7 +6,7 @@ use crate::application::caller::CallerContext;
 use crate::domain::entities::{AppId, OrganizationId, ProjectId, UserId};
 use crate::domain::errors::{DomainError, DomainResult};
 use crate::domain::value_objects::permission::{Permission, is_known_permission};
-use crate::domain::value_objects::role::name::RoleName;
+use crate::domain::value_objects::role::RoleName;
 use async_trait::async_trait;
 use derive_more::Constructor;
 use std::collections::{BTreeSet, HashMap};
