+++ crates/scylla-core/src/domain/value_objects/pipeline/mod.rs
@@ -1,11 +1,11 @@
 mod env;
+mod name;
 mod node_id;
-mod pipeline_name;
 mod step;
 mod working_dir;
 
 pub use env::*;
+pub use name::*;
 pub use node_id::*;
-pub use pipeline_name::*;
 pub use step::*;
 pub use working_dir::*;
