+++ crates/scylla-core/src/domain/value_objects/pipeline/name.rs
