+++ crates/scylla-core/src/domain/value_objects/role/mod.rs
@@ -1,1 +1,2 @@
 pub mod name;
+pub use name::RoleName;
