+++ crates/scylla-core/src/domain/value_objects/trigger/mod.rs
@@ -1,11 +1,11 @@
 mod cron_spec;
+mod name;
 mod trigger_input;
-mod trigger_name;
 mod trigger_source;
 mod webhook_spec;
 
 pub use cron_spec::*;
+pub use name::*;
 pub use trigger_input::*;
-pub use trigger_name::*;
 pub use trigger_source::*;
 pub use webhook_spec::*;
