+++ crates/scylla-core/src/domain/value_objects/trigger/name.rs
