--- examples/pipeline-verbose.json
@@ -1,61 +1,0 @@
-{
-  "name": "verbose-pipeline",
-  "projectId": "01ks7g0f8448j9rdng0p5e1498",
-  "nodes": [
-    {
-      "id": "setup",
-      "deps": [],
-      "command": "sh",
-      "args": [
-        "-c",
-        "echo '=== setup starting ==='; date; echo \"host: $(uname -a)\"; for i in $(seq 1 20); do echo \"[setup] preparing workspace step $i/20\"; done; echo 'warning: using default config' >&2; sleep 1; echo '=== setup done ==='"
-      ]
-    },
-    {
-      "id": "build",
-      "deps": [
-        "setup"
-      ],
-      "command": "sh",
-      "args": [
-        "-c",
-        "echo '=== build starting ==='; for i in $(seq 1 40); do echo \"[build] compiling module_$i.rs ... ok\"; if [ $((i % 10)) -eq 0 ]; then echo \"[build] progress: $i/40\"; sleep 1; fi; done; echo '=== build finished ==='"
-      ]
-    },
-    {
-      "id": "test",
-      "deps": [
-        "setup"
-      ],
-      "command": "sh",
-      "args": [
-        "-c",
-        "echo '=== test starting ==='; for i in $(seq 1 30); do echo \"[test] running case #$i\"; if [ $((i % 7)) -eq 0 ]; then echo \"[test] case #$i emitted a warning\" >&2; fi; done; sleep 1; echo '[test] 30 passed, 0 failed'; echo '=== test finished ==='"
-      ]
-    },
-    {
-      "id": "lint",
-      "deps": [
-        "setup"
-      ],
-      "command": "sh",
-      "args": [
-        "-c",
-        "echo '=== lint starting ==='; for f in main config handler router store; do for i in $(seq 1 6); do echo \"[lint] $f.rs:$i clean\"; done; done; echo '[lint] style nit on store.rs:4' >&2; echo '=== lint finished ==='"
-      ]
-    },
-    {
-      "id": "report",
-      "deps": [
-        "build",
-        "test",
-        "lint"
-      ],
-      "command": "sh",
-      "args": [
-        "-c",
-        "echo '=== report ==='; date; for i in $(seq 1 15); do echo \"[report] aggregating result $i/15\"; done; echo 'SUMMARY: build=ok test=ok lint=ok'; echo '=== pipeline complete ==='"
-      ]
-    }
-  ]
-}
